+++ tests/unit/test_parser_degruyter.py
@@ -27,25 +27,26 @@ class TestDeGruyterParser:
     def test_createLinks_isbn_match(self, testParser, mocker):
         parserMocks = mocker.patch.multiple(DeGruyterParser,
             generateS3Root=mocker.DEFAULT,
-            makeHeadQuery=mocker.DEFAULT,
             generatePDFLinks=mocker.DEFAULT,
             generateEpubLinks=mocker.DEFAULT
         )
 
         parserMocks['generateS3Root'].return_value = 'testRoot/'
-        parserMocks['makeHeadQuery'].return_value = (301, {'Location': 'degruyter.com/title/1'})
         parserMocks['generatePDFLinks'].return_value = ['pdf1', None]
         parserMocks['generateEpubLinks'].return_value = ['epub1', 'epub2']
 
-        testParser.uri = 'degruyter.com/9781234567890'
+        testParser.uri = 'degruyter.com/document/doi/10.000/9781234567890/html'
 
         testLinks = testParser.createLinks()
 
         assert testLinks == ['pdf1', 'epub1', 'epub2']
         parserMocks['generateS3Root'].assert_called_once()
-        parserMocks['makeHeadQuery'].assert_called_once_with('https://www.degruyter.com/view/books/9781234567890/9781234567890/9781234567890.xml')
-        parserMocks['generatePDFLinks'].assert_called_once_with('testRoot/')
-        parserMocks['generateEpubLinks'].assert_called_once_with('testRoot/')
+        parserMocks['generatePDFLinks'].assert_called_once_with(
+            'testRoot/', 'https://www.degruyter.com/document/doi/10.000/9781234567890/pdf'
+        )
+        parserMocks['generateEpubLinks'].assert_called_once_with(
+            'testRoot/', 'https://www.degruyter.com/document/doi/10.000/9781234567890/epub'
+        )
 
     def test_createLinks_title_match(self, testParser, mocker):
         parserMocks = mocker.patch.multiple(DeGruyterParser,
@@ -64,8 +65,12 @@ class TestDeGruyterParser:
 
         assert testLinks == ['pdf1', 'epub1', 'epub2']
         parserMocks['generateS3Root'].assert_called_once()
-        parserMocks['generatePDFLinks'].assert_called_once_with('testRoot/')
-        parserMocks['generateEpubLinks'].assert_called_once_with('testRoot/')
+        parserMocks['generatePDFLinks'].assert_called_once_with(
+            'testRoot/', 'https://www.degruyter.com/downloadpdf/title/1'
+        )
+        parserMocks['generateEpubLinks'].assert_called_once_with(
+            'testRoot/', 'https://www.degruyter.com/downloadepub/title/1'
+        )
 
     def test_createLinks_no_match(self, testParser, mocker):
         mockGenerate = mocker.patch.object(DeGruyterParser, 'generateS3Root')
@@ -85,34 +90,34 @@ class TestDeGruyterParser:
         mockHead.return_value = (200, 'mockHeader')
 
         testParser.uriIdentifier = 1
-        testLinks = testParser.generateEpubLinks('testRoot/')
+        testLinks = testParser.generateEpubLinks('testRoot/', 'epubSourceURI')
 
         assert testLinks == [
-            ('testRoot/epubs/degruyter/1/meta-inf/container.xml', {'reader': True}, 'application/epub+xml', None, None),
-            ('testRoot/epubs/degruyter/1.epub', {'download': True}, 'application/epub+zip', None, ('epubs/degruyter/1.epub', 'https://www.degruyter.com/downloadepub/title/1'))
+            ('testRoot/epubs/degruyter/1/META-INF/container.xml', {'reader': True}, 'application/epub+xml', None, None),
+            ('testRoot/epubs/degruyter/1.epub', {'download': True}, 'application/epub+zip', None, ('epubs/degruyter/1.epub', 'epubSourceURI'))
         ]
-        mockHead.assert_called_once_with('https://www.degruyter.com/downloadepub/title/1')
+        mockHead.assert_called_once_with('epubSourceURI')
 
     def test_generateEpubLinks_error(self, testParser, mocker):
         mockHead = mocker.patch.object(DeGruyterParser, 'makeHeadQuery')
         mockHead.return_value = (500, 'mockHeader')
 
         testParser.uriIdentifier = 1
-        testLinks = testParser.generateEpubLinks('testRoot/')
+        testLinks = testParser.generateEpubLinks('testRoot/', 'epubSourceURI')
 
         assert testLinks == [None]
-        mockHead.assert_called_once_with('https://www.degruyter.com/downloadepub/title/1')
+        mockHead.assert_called_once_with('epubSourceURI')
 
     def test_generatePDFLinks(self, testParser, mocker):
         mockGenerate = mocker.patch.object(DeGruyterParser, 'generateManifest')
         mockGenerate.return_value = 'testManifestJSON'
 
         testParser.uriIdentifier = 1
-        testLinks = testParser.generatePDFLinks('testRoot/')
+        testLinks = testParser.generatePDFLinks('testRoot/', 'pdfSourceURI')
 
         assert testLinks == [
             ('testRoot/manifests/degruyter/1.json', {'reader': True}, 'application/pdf+json', ('manifests/degruyter/1.json', 'testManifestJSON'), None),
-            ('https://www.degruyter.com/downloadpdf/title/1', {'download': True}, 'application/pdf', None, None)
+            ('pdfSourceURI', {'download': True}, 'application/pdf', None, None)
         ]
 
     def test_generateManifest(self, testParser, mocker):
@@ -139,5 +144,5 @@ class TestDeGruyterParser:
 
         assert DeGruyterParser.makeHeadQuery('testURI') == (200, 'testHeaders')
         mockHead.assert_called_once_with(
-            'testURI', timeout=5, headers={'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_5)'}
+            'testURI', timeout=15, headers={'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_5)'}
         )